Tests: MonadCombine->Alternative, add missing ones
